CA-185079: set-vg-info always called with local_allocator path
In the past we were rewriting the vg-info file with the local_allocator
path set only when the local_allocator was started.
Given some changes in xenvm it is not only safe but mandatory
to make sure the local_allocator path is always written down.

Our wrapper, setvginfo, now does not accept any more an additional
parameter but it will always pass the local_allocator path
to set-vg-info, regarldess of the daemon running or not.

Signed-off-by: Germano Percossi <germano.percossi@citrix.com>