CA-192760: use noop IO scheduler for multipath and non-local devices
We need to ensure we use noop IO scheduler for multipath devices
and non local devices.

We were used to do that in the storage manager but for devices
going offline and added back this is not possible, becasue
SM is not running all the time.

This patch ensures the noop scheduler is set for specific devices
and this setting is preserved even when SM is not running
or when SM is not managing multipath (a potentially
upcoming feature).

Whit this rule, we use noop for any multipath device or any
block device that is not on the ATA bus (a very weak decision
rule to get an idea of locally attached devices).

Signed-off-by: Germano Percossi <germano.percossi@citrix.com>