squash on commit 12ba5cb and keep the following commit message
CA-192760: use noop IO scheduler for specific device types

We need to ensure that a noop scheduler is used for the following
devices:
- non rotational disks (e.g. SSDs)
- multipath devices (we assume they have a clever controller if
  many paths exist)
- remote arrays

In the latter case we are not sure the chosen rule is always
doing the right thing.

We were used to do that in the storage manager but for devices
going offline and added back this is not possible, because
SM is not running all the time.

This patch ensures the noop scheduler is set for specific devices
and this setting is preserved even when SM is not running
or when SM is not explicitly managing multipath (a potentially
upcoming feature).

Signed-off-by: Germano Percossi <germano.percossi@citrix.com>