Fix LVM unit tests.
The UT depend on a specific path for lvcreate binaries, however
the path actually used is distro dependent.  On Debian distros,
lvcreate is in /sbin.  SM correctly auto detects the location,
but the UT were expecting it in /usr/sbin.

Fix the UT so they use the detected path for the executables.