CP-16977: Add dependecy on xenserver-lvm2
From: Germano Percossi <germano.percossi@citrix.com>

SM relies on functionality or bugfixes provided by
the custom xenserver-lvm2, so this needs to
be enforced.

If we shipped with a "vanilla" lvm2
we could get rid of this dependency but it could still
be useful even with an unpatched lvm2 (e.g. to make
sure a specific version of lvm2 is used).

Signed-off-by: Germano Percossi <germano.percossi@citrix.com>