CP-16736: Add dependecy on xenserver-multipath
From: Germano Percossi <germano.percossi@citrix.com>

SM relies on functionality or bugfixes provided by
the custom xenserver-multipath, so this needs to
be enforced.

If we shipped with a "vanilla" device-mapper-multipath
we could get rid of this dependency but it could still
be useful even with an unpatched multipath (e.g. to make
sure a specific version of multipath is used).

Signed-off-by: Germano Percossi <germano.percossi@citrix.com>
Reviewed-by: Mark Syms <mark.syms@citrix.com>