CA-269654: Make tapdisk-unpause CBT aware
When a disk is attached by the user/XAPI, the VDI class checks if CBT is
enabled on the disk and passes the right parameters to tap-unpause and
tap-refresh. A GC thread is CBT-unaware and reattached lead VDIs without
the CBT log file after a leaf coalesce. The patch moves the onus of
being CBT aware to tapdisk unpause so that the right CBT log file can be
attached irrespective of the caller

Signed-off-by: Chandrika Srinivasan <chandrika.srinivasan@citrix.com>
Reviewed-by: Mark Syms <mark.syms@citrix.com>