CA-270011: Handle snapshotting a snapshot in a CBT chain
Snapshotting a snapshot is rare but possible. When this happens on a
snapshot that is part of the CBT chain, insert the new child of the
snapshot into the CBT chain.

We are trying to solve the case of snapshotting a snapshot in a CBT chain.
Consider the following CBT chain P1<->P0<->V->NULL.
Under normal circumstances, V is snapshotted and we do the following

1) Rename V to S (snapshot) 
2) Find parent of S (P0 in this case) and create P0<->S links
3) Create new V (child of V is set to NULL, since this is usually the leaf) 
4) Create S<->V links
5) Final chain looks like P1<->P0<->S<->V->NULL

This is very specific to a leaf being snapshotted and this needs to be
generalised for any VDI in the chain being snapshotted. Consider the scenario
where P0 is being snapshotted. The process needs to change to

1) Rename P0 to S (snapshot) 
2) Find parent of S (P1 in this case) and create P1<->S links
3) Create new P0 (child of P0 is still set to NULL at the time of creation)
4) Detect if the original P0 (now S) had a child (V in this case)
5) Link P0 to V
6) Create S<->P0 links
7) Final chain looks like P1<->S<->P0<->V->NULL

Signed-off-by: Chandrika Srinivasan <chandrika.srinivasan@citrix.com>