CP-22423: Convert XenError into an exception factory
Change XenError just enough to make it produce SROSExceptions
instead or raising them in the constructor.

Because XenError was already raising exceptions in the
constructor, writing
 raise XenError
was redudant even though the right thing to write to make
intentions clear.

One test case was exploiting the fact that 'raise' did not
need to be typed and the test has beeen fixed

Signed-off-by: Germano Percossi <germano.percossi@citrix.com>
Reviewed-by: Chandrika Srinivasan <chandrika.srinivasan@citrix.com>
Reviewed-by: Mark Syms <mark.syms@citrix.com>