CA-274584: Provide SM service to take care of reboot issues with iscsi
Upon reboot/shutdown Xapi does not call pbd-unplug to gracefully
detach the SRs.

In this specific case, iscsid has been started by SM and iscsi
sessions have been logged in by SM.
Under some circumstances, iscsid will refuse to shutdown because there
are some iscsi sessions on the system that have not been closed: SM
did not get any chance to do it.

This service introduces a dependency on iscsi-shutdown service.

In order for this to work, the spec file (or else) needs to contain
the code to enable this service.

This should fix CA-272842, too.

Signed-off-by: Germano Percossi <germano.percossi@citrix.com>
Reviewed-by: Mark Syms <mark.syms@citrix.com>