CA-276751: clean-up wwids at start-up
This is another issue due to the fact that Xapi does not unplug
the SRs upon reboot.

SM is responsible for the multipath wwids file but if not called
to detach the SRs it does not have the chance to clean it.

This addition to the service clean-up every wwid that is not in use
during reboot.

Because it is very unlikely that something is already in use during
reboot (apart from the root disk), this prevents users from adding
their own wwids to mark some devices as special.
However, this behavious is not discouraged anyway, because SM needs
to have full control over multipath.

Signed-off-by: Germano Percossi <germano.percossi@citrix.com>
Reviewed-by: Mark Syms <mark.syms@citrix.com>