CA-288222: deactivate LV on slaves before changing the LVM layout
When creating snapshot for a LVHD attached on a slave host, we firstly
change the LV  on the master and then refresh the LV on the slave. As
part of the process, we'll run "lvchange -an" on the slave and expect
it would tear down the existing DM (so that we can create new one based
on the new configuration). At the point when we do this, the LV's
configuration has already been changed (by master) hence differs from
the existing DM at runtime. It's unclear, whether in this case, the
lvchange will still tear down the (now diverted) DM instance. It
shouldn't be a suspire if lvchange refuses to do so for safety
consideration, or simply fails to do so because of the existence of some
assertion in place. Anyway, it looks like an undefined behaviour which
is implementation dependent.

To avoid uncertainty, this patch shift the deactivateNoRefcount step
forward, to before any LV changes taking place on the master, where the
LV configure and DM setup still match each other.

Signed-off-by: Zheng Li <zheng.li3@citrix.com>