CA-295775 Fix handling of multipath events
Sending multipath UDEV events via a systemd service into atd does not
work very well at getting them processed; systemd drops attempts to
start any service which is already in the process of starting (which
loses information in the case of an event storm) while atd is a batch
processor designed to do things "at some vague time in the future which
could turn out to be never if the system load is higher than some
number".

A better solution is to use neither, and when we are scanning for
updates because of a UDEV event, use a simple wait/run lock pair to
rate-limit scans while ensuring that every event is followed by a scan
as soon as possible so its new state is detected.

We can now remove the 2 second sleep in drivers/mpathcount.py

In addition, the match_pathup function in drivers/mpathcount.py was not
matching the 'faulty' status in the multipath output and thus was not
updating the status when it changed. (Cosmetic: removed a tab from a log
line, log why the db update failed if it did fail).

In drivers/mpath_cli.py invoke multipathd by its full path because we
know it. Do not rely on PATH search for things like this; udev does
not supply the PATH we want and we should avoid requiring it.