CA-286144: udev script calling xe pusb-scan on boot timing out when XAPI is not up yet
This fix is trying to address the issue that when xapi is not yet ready,
the udev rules is triggered and possible hang.

The fix creates a new Python scripts as daemon service, waiting on a pipe.
The udev rules will invoke another script (the same one but with different
mode) to trigger the daemon script to try to trigger xapi usb-scan.

The implementation assumes the server side to be run as systemd daemon, and
it waits on a named pipe, which when the udev rules are triggered, the wrapper
script will write hence wake up the server and the server will further invoke
the script to do the actual work as child.

Signed-off-by: YarsinCitrix <yarsin.he@citrix.com>