CA-262506: Remove sec=ntlm from SMBSR mount option
With pre-4.14 kernel, if the security type specified using "sec=" mount option
is not supported, the SMB2 session setup code changes the security type
to RawNTLMSSP. This behaviour has been changed in 4.14.y, where it would now
fail the mount and return an error, if "sec=" mount option is not supported.

SMB2 session setup code now uses the global security flags
(/proc/fs/cifs/SecurityFlags) to select the security method to be used when
no security method is specified.

Signed-off-by: Mark Syms <mark.syms@citrix.com>
Signed-off-by: Anoob Soman <anoob.soman@citrix.com>