CP-29778: detect encrypted VHDs during SR scan
For each VHD that we find, we try to get the encryption key hash using
vhd-util. We add this to the sm_config_override map of the VDI, which
will cause the VDI's sm_config in xapi's DB to be updated as necessary
at a later stage during SR scan.

Thus for new encrypted VHDs, the key_hash sm_config field will be added
to xapi's DB during SR scan, and xapi and SM will know that this is an
encrypted VDI and which key it requires.

Signed-off-by: Gabor Igloi <gabor.igloi@citrix.com>