CA-299722 Adjust timeout and retrans NFS mount options
Keep the NFS timeo mount option in deciseconds always. Various parts of
the code appear to have tried to have it in seconds, except for one
place which apparently thought it was in centiseconds. This
inconsistency resulted in a timeout/retrans combination of 1000
deciseconds and 12 retries, which would mean it would take over 1.5
hours for an NFS operation to fail. This hides failures in the NFS
layer.

Adjust everywhere know that the timeo value is in deciseconds. Also,
since tapdisk will retry EIO, it makes sense to make the retry period
shorter and not to try so many times, so lower the defaults.