CA-293816: convert OSError to CommandException
It was already the case for all OSError to be converted to
CommandException here, but this doesn't hold for repeated EIO,
where the OSError will still be raised in the end.

However, majority of the exception catching code that wraps
around ioretry will only catch CommandException, and some will
try to get its "code" field which doesn't exist for OSError.

It seems to be the intention of the original code that all OSError
should be converted to CommandException here.

Signed-off-by: Zheng Li <zheng.li3@citrix.com>