CA-309979 Fix Storage Manager initialisation
Now that we have xapi-init-complete.target, both mpathcount.service and
usb-scan.service can wait until this point as there is no point to
either of them until xapi can be notified of their results.

In addition, ExecStartPost can be used in both cases to kick them once
as soon as they come up, which makes the mpathroot init script obsolete.

The sm-multipath init script is likewise incorrect and its reliance on
systemd-sysv-generator coupled with its LSB header cause it to come up
*after* network-online.target which is too late. Since it that there are
non-init uses of this script, leave it alone for now but do not enable
it (and disable it on upgrade). Replace its service function with the
new oneshot sm-mpath-root.service which calls a simplified version.

We should find and replace the non-init uses of
/etc/rc.d/init.d/sm-multipath so that we can stop providing it entirely
and eventually remove all use of INITDIR from Storage Manager

Signed-off-by: Tim Smith <tim.smith@citrix.com>