CA-323702: fcntl locks are per process, so need to reference count.
As fcntl locks are per-process it is not an error if a lock is acquired
that is already held, however the release is not reference counted by
the system. So, if a lock is taken by an API and subsequently another
method in the call chain takes the same lock it must not be released
until the API call completes.

To ensure this behaviour make Lock objects unique by name and namespace
so that subsequent calls obtain the same object reference. internally
to the object track how many calls to acquire and release have been made
and only release the underlying lock when all the acquires have had matching
release calls.

Signed-off-by: Mark Syms <mark.syms@citrix.com>