CA-334762: Fix VDI import with storage driver domains
With storage driver domains a VDI/VM import would fail:
Xenops_interface.Xenopsd_error([S(Device_detach_rejected);[S(VBD);S(xvdn);S(16 Device in use; refusing to close)]])

Using blktrace/blkparse and with-vdi has shown that `udev` is accessing
the device and launching various other programs to poke at it:
is it a cdrom? is it multipath? etc.

We do not want any of that processing, the sole purpose of these devices
plugged into Dom0 is to allow operations like import/export to work.

So modify the udev rules to ignore these devices as much as possible.
Udev no longer has the feature to ignore a device or stop processing rules,
but we can define various ENV vars that are meant to skip various rules.
The `nowatch` option should also prevent generating change events.

This depends on having a new enough systemd, or a backported version of
https://github.com/systemd/systemd/commit/542127ea96de4b0e22fbb0fc872d80406e068e36

Put the ignore rules into a section of its own to allow td*/nbd* devices
to benefit from these rules in a followup commit.

Signed-off-by: Edwin Török <edvin.torok@citrix.com>