CA-368769: extend the timeout on tap-ctl close
In XSI-620 the storage manager scripts were blocked by a tapdisk
process which was hung totally and blocked for over an hour. A
timout was added to this close operation to unblock things but the
timeout is only on the IPC and so breaks the IPC link if the timeout
is reached. Which is fine if the remote process is entirely wedged,
however when a close is requested and there are pending IO requests,
possibly during a migrate, the close operation can take significantly
longer than the currently permitted 30 seconds, 50 seconds observed
in failing tests. So, extend the timeout such that this scenario
should be handled successfully whilst not reintroducing the issues
seen in XSI-620.

Signed-off-by: Mark Syms <mark.syms@citrix.com>