CP-41402: check for NBD socket connections before closing parent
Historically SM has checked for tapdisk processes which have the
/dev/xen/blktap-2/tapdev<minor> block device open for the parent's
minor device id and closed the parent tapdisk when there are no
remaining children. This condition doesn't hold true when blktap3 is
updated to use the NBD server Unix socket when connecting to the
parent process in cache mode. Update the check to also look for
tapdisk processes which have a connected, state 03, Unix socket for
the nbd<pid>.<minor> socket of the parent minor id, pid is ignored as
only one process can have the referenced minor.

Signed-off-by: Mark Syms <mark.syms@citrix.com>