mk/sm.spec.in: remove uninstall operations now unrelated to this package
The sm package doesn't create /etc/lvm/master/lvm.conf anymore, nor
alter the main /etc/lvm/lvm.conf, anymore, so it's not its role anymore
either to restore the configuration when uninstalled.

Also remove the else part, as calling true is not needed in the case of
a RPM upgrade: nothing is executed before (the systemd preun macros only
do anything on an uninstall).

Signed-off-by: Samuel Verschelde <stormi-xcp@ylix.fr>