CP-27709: filter error messages about ioctl not supported in trim
On a LUN which has provisioning_mode = full the blkdiscard ioctl will
usually fail which is not entirely unexpected as the primary use case
for blkdiscard here is to release unused blocks on a thinly
provisioned LUN so that a SAN may reassign them to a different LUN if
required. Leave the code making the call as it is possible that some
SANs may still accept a blkdiscard on a fully provisioned LUN and
usefully do something with the request, but do not propagate an error
back to the user as it just creates confusion.

Signed-off-by: Mark Syms <mark.syms@citrix.com>