CA-376740: smb: Fix snapshot with newer kernels
Newer kernels defer closes by default but this interferes with use of
hard links since with SMB (or maybe the underlying server filesystem), a
hard link cannot be opened if the target is also open, which it may be
temporarily with deferred close. Disable using the mount param available
since Linux 6.0.

This can be reproduced with the following Python snippet. When deferred
close is enabled, the final open fails with EINVAL.

    fd = os.open('test.vhd', os.O_WRONLY|os.O_CREAT)
    os.write(fd, b'foo')
    os.close(fd)

    fd = os.open('test.vhd', os.O_RDONLY|os.O_DIRECT)
    os.close(fd)

    os.link('test.vhd', 'new.vhd')
    newfd = os.open('new.vhd', os.O_RDONLY|os.O_DIRECT)

Signed-off-by: Ross Lagerwall <ross.lagerwall@citrix.com>