CA-379329: check for missing iSCSI sessions and reconnect
If, when an iSCSI SR is attached, some of the paths are not reachable
they will not be subsequently connected should the connectivity issues
be resolved, without detaching the SR or rebooting the host. Neither
of these options give a good experience for users of the software. Add
a systemd timer which will, periodically (default 10 minutes), ask
each SR to do SR specific "health" checks. Currently only a checker
for the iSCSI sessions is present but more can be added to this
pattern in the future as needs arise.

Signed-off-by: Mark Syms <mark.syms@citrix.com>