CA-379287 Cope with fs-encoded XMLRPC request on command line
When SRCommand parses an XMLRPC request, convert is back to the byte
string that the python process would have received from the command line
before trying to parse it. This is necessary because it will have been
encoded using UTF-8, and where characters need to be represented as
multi-byte sequences these may show up as "surrogate escapes"
(cf PEP 383).

Signed-off-by: Robin Newton <robin.newton@cloud.com>