CP-39600: Rework LVM locking to use fair lock queue
This is intended to prevent a storm of later LVM commands from usurping
the lock that an earlier command is waiting for, just because they
happened to manage to start waiting before the original holder was done.

In some pathalogical circumstances, this has been known to result in a
commant taking many minutes to get the lock.

Signed-off-by: Tim Smith <tim.smith@cloud.com>