CA-386281 CIFS username can be omitted in ISO SR
ISOSR falling over when no username is given is a regression that looks
to have been introduced in commit
cbd7226b012d145b52ef50c45c69987d4050b12a when the use of a credentials
file was removed.

The ISOSR unit tests were already covered cases of not having
credentials - as evinced by having "guest" in the expected mount
options. However, they were doing this in a rather confused way that hid
the bug by supplying a username regardless. Now only the credentials
tests do this.

Signed-off-by: Robin Newton <robin.newton@cloud.com>