CA-390490: remove override of blktap request timeout
When this code was orginally added there was no default value for the
nfs_timeout and so the timeout would only have been used if
1. The SR type was NFS
2. The SR had a overriden setting for nfs-timeout in the other-config

Since that time the calculation of the NFS client timeout has been
changed considerably and is set with a balance between timeo and
retrans values. In particular the get_nfs_timeout method will now
always return a non-zero value.  This results in the SM always
applying a request timeout to blktap of 50 seconds (unless the timeo
value has been set in the SR device config), which is considerably
lower than the default request timeout in blktap itself which is 120
seconds.

NFS protocol timeouts are also no longer considered to be retryable
errors within blktap so tying the timeout value to the NFS timeout,
even if it were correctly calculated taking into account the retrans
value, is effectively comparing two unrelated things.

Signed-off-by: Mark Syms <mark.syms@citrix.com>