CP-45750 Allow for alternative local storage SR types
storage-init no longer silently ignores unrecognized SR types, but
instead handles them in the same way as 'lvm' and 'ext' (the handling of
which has now been unified).

This presupposes that an SR type specified in first-boot configuration
will take a single configuration parameter - 'device' - which is a
comma-separated list of partitions - which is what 'lvm' and 'ext' do.

Signed-off-by: Robin Newton <robin.newton@cloud.com>