CA-387861 Introduce fair locking subsystem (#683)
None of the file-based locking systems available are "fair", as in,
there is no guarantee that the first waiter will not be usurped when the
lock is released by later waiters, thus suffering from lock starvation.
This has been something of a problem across various attempts to wrap
access to LVM, and indeed it turns out that multipath commands also need
to be serialised on some of the same resources.

This adds a small templated systemd service which does nothing but
accept one connection at a time on a UNIX Domain socket, with a fairly
long accept queue. The service is started when needed (thus every
process using this mechanism needs permission to start the service), and
the "Lock" is acquired by connecting to the socket. connect() on UNIX
Domain sockets does not time out, and waiting connections are held in a
queue, providing fairness.

The python3 context manager is provided via a metaclass which checks the
arguments and allows only one of each named Fairlock object to exist, so
that attempts to take a lock when it is already held can raise an
exception. This is not thread-safe but could probably be made so.

Signed-off-by: Tim Smith <tim.smith@cloud.com>
Co-authored-by: Tim Smith <tim.smith@cloud.com>