CA-392823: ensure no device mapper conflicts in LVHDSR detach
Within the detach operation checks are made that none of the device
mapper devices for the SR volume group have opern handles. The falls
false if operations on other LVM srs happen to call a device mapper
operation such as `lvs` or `vgs` at the point the check is made as
these operations can result in all of the devices managed by the
device mapper subsystem being accessed as part of the scan. this then
gives a false response for the open file handles check.

Signed-off-by: Mark Syms <mark.syms@cloud.com>