CA-390937: don't clean caches until we have our locks
Doing the cleanupCache before acquiring the locks can result in a
failure, in particular as part of XenCert where an sr_detach operation
attempts to abort GC but the GC is not registered as running and then
the umount of NFS fails as the cache cleanup still has the directory
open.

Also check for the SR still being attached after the active lock has
been acquired.

Signed-off-by: Mark Syms <mark.syms@cloud.com>