CP-49720 Move LOCK_TYPE_RUNNING from cleanup.py to lock.py
In the process, rename LOCK_TYPE_RUNNING to LOCK_TYPE_GC_RUNNING to
describe *what* is running when it's running. This matches
LOCK_TYPE_GC_ACTIVE which remains in cleanup.py because it is only used
there.

Rename the global variables in cleanup.py to describe *what* is
Active/Running.

Signed-off-by: Tim Smith <tim.smith@cloud.com>