CP-49775 convert SMGC to systemd service
Add a templated oneshot service and use systemctl to start it. The
systemctl command is invoked with "--no-wait" so that it returns
immediately because the "oneshot" nature of the service would otherwise
wait until the command completes.

Note that because the service is a oneshot with RemainAfterExit=no (so
we can multiply start it), it will go from "activating" to "dead"
without ever appearing to be "running".

Signed-off-by: Tim Smith <tim.smith@cloud.com>