CA-393194: Fix pvremove failure
In lvmohba test, the test failed at
`SR.destroy(sr_ref) -> removeVG -> pvremove ` with the
lvm error message "No devices to process".
The root cause is that the PV which was to be removed
was not really used as PV when it comes to multipathed devices.
The LVM official document says:
"Since each LUN has multiple device nodes in /dev 
that point to the same underlying data,
they all contain the same LVM metadata
and thus LVM commands will find the same metadata
multiple times and report them as duplicates.
These duplicate messages are only warnings
and do not mean the LVM operation has failed.
Rather, they are alerting the user that only one of the devices
has been used as a physical volume and the others are being ignored." 
Please refer to https://docs.redhat.com/en/documentation/
red_hat_enterprise_linux/6/html/logical_volume_manager_administration
/duplicate_pv_multipath#duplicate_pv_multipath

This fix is to find the real PV in a VG before removing the VG.

Signed-off-by: Stephen Cheng <stephen.cheng@cloud.com>