CA-395554 Improve fairlocking reliability
Under some circumstances (when the listen queue is short enough) a
process's blocking connect() to a UNIX-domain stream socket can return
success even when the server side has not yet called accept().

To work around this, we have the server portion write a small fixed blob
of data to each new connection and require the client to read it before
it considers itself to have the lock; merely being connected
successfully is no longer sufficient.

Signed-off-by: Tim Smith <tim.smith@cloud.com>