CA-398958 Cope with concurrent read-only activations
It should be possible to activate a VDI on two different hosts in a
pool simultaneously, provided read-only access is all that it required.
So, if one host tries to activate in a brief window when another host
has put the transient "activating" key on the VDI, it should be allowed
to retry.

Signed-off-by: Robin Newton <robin.newton@cloud.com>