CA-400106: disable leaf coalesce with VDI snapshot secondary
When the secondary/mirror option is provided to the vdi_snasphot
operation we do not want any leaf coalesce garbage collection taking
place. Unfortunately xapi will delete the snapshot once it's finished
copying the new parent to the destination SR in SXM and whilst the
mirroring operation is still live. This will kick the GC process and
if the GC gets to the point of performing the leaf coalesce before the
migrate completes will result in a failure of the mirror operation. So
block the leaf-coalesce operation in this case. The block will be
removed if another snapshot were to be taken but in the case of SXM
the next expected operation is a vdi_delete of the leaf which will
clear the entire chain.

Signed-off-by: Mark Syms <mark.syms@cloud.com>