CA-397084: SR scan tries to deactivate LV in use by tapdisk
Back when this code was written we didn't have leaf coalesce doing
snapshot coalesace operations, and so for performance reasons the scan
code didn't reference count LVs that it activated on the assuption
that nothing else would be in a position to start using the LV in the
meantime. This assumption is no longer valid and induces scan failures
when a new VDI is detected and the scan tries to deactivate it whilst
it is open and in use by tapdisk. Move the scan code to use the
lvactivator so that reference counts will be taken and
honoured. Additionally, it's possible that more than one VDI might
appear so don't treat this as a singular event and track all VDIs that
might have been activated and dereference them at the end.

Signed-off-by: Mark Syms <mark.syms@cloud.com>