CP-52852: add handler for xmlrpc ProtocolError
It is possible that a XenAPI call to xapi will result in an xmlrpc
ProtocolError being returned, typically a 500 Internal Error. As it is
not practical to wrap all the calls to xapi via the xenapi session in
a specific exception handler add one into the outer run method to
ensure that it is converted to an error code and not an unknown error.

Signed-off-by: Mark Syms <mark.syms@cloud.com>